requirements.txt
# requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
pyodbc==5.0.1
bcrypt==4.0.1
dash==2.14.2
plotly==5.19.0
pandas==2.2.1
kaleido==0.2.1  # necessário para exportar PDF com Plotly
